Data Validation

+++ api/Controllers/StockController.cs
@@ -27,6 +27,8 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var stocks = await _stockRepo.GetAllAsync();
             var stockDto = stocks.Select(s => s.ToStockDto());
             return Ok(stocks);
@@ -37,18 +39,22 @@ namespace api.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> GetOne([FromRoute] string id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var stock = await _stockRepo.GetOneAsync(id);
-            if (stock==null) { return NotFound(); }
+            if (stock == null) { return NotFound(); }
             return Ok(stock.ToStockDto());
         }
 
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var stockModel = stockDto.ToStockFromCreateDTO();
             await _stockRepo.CreateAsync(stockModel);
 
-            return CreatedAtAction(nameof(GetOne), new {id = stockModel.Id}, stockModel.ToStockDto());
+            return CreatedAtAction(nameof(GetOne), new { id = stockModel.Id }, stockModel.ToStockDto());
         }
 
 
@@ -57,9 +63,11 @@ namespace api.Controllers
         [Route("{id}")]
         public async Task<IActionResult> Update([FromRoute] string id, [FromBody] UpdateStockRequestDto updateDto)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
 
-            if(stockModel == null)
+            if (stockModel == null)
             {
                 return NotFound();
             }
@@ -73,10 +81,12 @@ namespace api.Controllers
         [Route("{id}")]
         public async Task<IActionResult> Delete([FromRoute] string id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var guid = Guid.Parse(id);
             var stockModel = await _stockRepo.DeleteAsync(id);
 
-            if (stockModel==null) {return NotFound();}
+            if (stockModel == null) { return NotFound(); }
 
             return NoContent();
         }
